+++ ios/RNDeviceInfo/RNDeviceInfo.h
@@ -9,7 +9,11 @@
 #import <UIKit/UIKit.h>
 #import <sys/utsname.h>
 
+#if __has_include(<React/RCTAssert.h>)
 #import <React/RCTBridgeModule.h>
+#else
+#import "RCTBridgeModule.h"
+#endif
 
 @interface RNDeviceInfo : NSObject <RCTBridgeModule>